requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
idna==2.5
lxml==3.7.3
mss==2.0.22
numpy==1.12.1
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
protobuf==3.3.0
requests==2.17.3
robobrowser==0.5.3
six==1.10.0
tensorflow-gpu==1.1.0
urllib3==1.21.1
Werkzeug==0.12.2
